server/pyproject.toml
[tool.poetry]
name = "communex"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = { version = "^0.22.0", extras = ["standard"] }
pydantic = "^2"
pydantic-settings = "^2"
yarl = "^1.9.2"
ujson = "^5.8.0"
SQLAlchemy = { version = "^2.0.18", extras = ["asyncio"] }
alembic = "^1.11.1"
redis = { version = "^4.6.0", extras = ["hiredis"] }
httptools = "^0.6.0"
aio-pika = "^9.1.4"
taskiq = "^0"
taskiq-fastapi = "^0"
taskiq-redis = "^0"
taskiq-aio-pika = "^0"
sqladmin = "^0.16.0"
python-socketio = "^5.10.0"
asyncpg = "^0.28.0"
tortoise-orm = "^0.19.3"
aerich = "^0.7.1"


[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~4.0.1"
mypy = "^1.1.1"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^22.12.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
fakeredis = "^2.5.0"
httpx = "^0.23.3"
taskiq = { version = "^0", extras = ["reload"] }

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["TaskiqTemplateProject", ]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true
# Remove this and add `types-redis`
# when the issue https://github.com/python/typeshed/issues/8242 is resolved.
[[tool.mypy.overrides]]
module = [
    'redis.asyncio'
]
ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "TASKIQTEMPLATEPROJECT_ENVIRONMENT=pytest",
    "TASKIQTEMPLATEPROJECT_DB_BASE=TaskiqTemplateProject_test",
]
[fastapi-template.options]
project_name = "TaskiqTemplateProject"
api_type = "rest"
enable_redis = "True"
enable_rmq = "True"
ci_type = "none"
enable_migrations = "True"
enable_taskiq = "True"
enable_kube = "None"
kube_name = "TaskiqTemplateProject"
enable_routers = "None"
enable_kafka = "None"
enable_loguru = "None"
traefik_labels = "None"
add_dummy = "True"
orm = "sqlalchemy"
self_hosted_swagger = "None"
prometheus_enabled = "None"
sentry_enabled = "None"
otlp_enabled = "None"
pydanticv1 = "None"
gunicorn = "None"
add_users = "None"
cookie_auth = "None"
jwt_auth = "None"
[tool.aerich]
tortoise_orm = "src.db.config.TORTOISE_CONFIG"
location = "./src/db/migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"